+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to
 
 ## Unreleased
 
+### Changed
+
+- Added option to allow duplicate types in `getSortedJupiterOneTypes`
+- Changed `j1-integration document` command to allow duplicate types
+
 ## 8.28.1 - 2022-11-9
 
 ### Changed